+++ server/src/models/Review.js
@@ -15,6 +15,16 @@ class Review extends Model {
         }
     }
 
+    get votedStatus(){
+        let voted
+        if (this.userId === user.id) {
+            voted = true
+        } else {
+            voted = false
+        } 
+        return voted
+    }
+
     static get relationMappings() {
         const { User, Activity, Rating } = require("./index.js")
